Add view option "List" to Documentation
Initial View supports two different options, "ListWeek" shows a list starting on the first day of the week, whereas "List" shows a list beginning at the current day. I had to read through a lot of gits to find that option, it'd be nice if it were explained in the doc.